pyproject.toml
[tool.poetry]
name = "django-socketio-chat"
version = "0.1.1"
description = "Django SocketIO Chat using python-socketio"
authors = ["kai <thuan.dv0@gmail.com>"]
repository = "https://github.com/ThuanD/django-socketio-chat/"
readme = "README.md"
packages = [
    { include = "django_socketio_chat" },
]

[tool.poetry.dependencies]
python = "^3.10"
django-allauth-simple-ui = "^0.1.2"
djangorestframework-simplejwt = "^5.3.1"
python-socketio = "^5.11.4"
gunicorn = "^23.0.0"
eventlet = "^0.37.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
